requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
frozenlist==1.5.0
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.2.0
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.0
urllib3==2.3.0
win32_setctime==1.2.0
yarl==1.18.3
